Tests finally passing.
Almost lost my fucking mind on the metaclasses there, but
somehow pulled out of the dive by doing exactly what I'd
been doing for hours, except magically it's right now.

No more commits that break the core tests.
This is the core of how to build an all2vec system.
Everything else has to go on top of this, without
breaking the core.