Significant change to Type.project leading to various improvements.
This commit contains changes whose necessity became apparent
after the recent implementation of "ordinal bases" as a
conceptual foundation for decimal and binary thoughts.
The biggest resulting change is that 'project' for all
attributes now finds the most similar member of the attribute's
basis, rather than performing a call to slow.attention_l1 as
before. The difference is that we're now using effectively
the argmax of pre_attention_l1's value, rather than the
similarity-weighted average of every member of the basis.
Long term, this may prove not to be the most general solution,
and types should be able to swap out different methods of
projecting onto themselves and inverting thoughts with respect
to themselves. But for now, the improvement is seriously
impressive. Training objects and types is now much quicker
and much more reliable. It was hard to see the flaw until
the highly non-orthogonal nature of the "ordinal basis"
concept demonstrated how difficult training even simple
objects became if we used the similarity-weighted average
over the basis, rather than the argmax. The map remains
learnable both insofar as it depends on the position of
the object and the positions of all the values in the
attribute's basis. There is no reason to muddy the waters
by mixing in bad answers. Great progress today. ^_^